+++ exporter/tanzuobservabilityexporter/factory.go
@@ -20,7 +20,6 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/resourcetotelemetry"
@@ -39,18 +38,10 @@ func NewFactory() component.ExporterFactory {
 }
 
 func createDefaultConfig() config.Exporter {
-	tracesCfg := TracesConfig{
-		HTTPClientSettings: confighttp.HTTPClientSettings{Endpoint: "http://localhost:30001"},
-	}
-	metricsCfg := MetricsConfig{
-		HTTPClientSettings: confighttp.HTTPClientSettings{Endpoint: "http://localhost:2878"},
-	}
 	return &Config{
 		ExporterSettings: config.NewExporterSettings(config.NewComponentID(exporterType)),
 		QueueSettings:    exporterhelper.NewDefaultQueueSettings(),
 		RetrySettings:    exporterhelper.NewDefaultRetrySettings(),
-		Traces:           tracesCfg,
-		Metrics:          metricsCfg,
 	}
 }
 
@@ -61,12 +52,17 @@ func createTracesExporter(
 	set component.ExporterCreateSettings,
 	cfg config.Exporter,
 ) (component.TracesExporter, error) {
+	tobsCfg, ok := cfg.(*Config)
+	if !ok {
+		return nil, fmt.Errorf("invalid config: %#v", cfg)
+	}
+	if err := tobsCfg.sanitize(); err != nil {
+		return nil, err
+	}
 	exp, err := newTracesExporter(set, cfg)
 	if err != nil {
 		return nil, err
 	}
-
-	tobsCfg := cfg.(*Config)
 
 	return exporterhelper.NewTracesExporter(
 		cfg,
@@ -83,12 +79,17 @@ func createMetricsExporter(
 	set component.ExporterCreateSettings,
 	cfg config.Exporter,
 ) (component.MetricsExporter, error) {
+	tobsCfg, ok := cfg.(*Config)
+	if !ok {
+		return nil, fmt.Errorf("invalid config: %#v", cfg)
+	}
+	if err := tobsCfg.sanitize(); err != nil {
+		return nil, err
+	}
 	exp, err := newMetricsExporter(set, cfg, createMetricsConsumer)
 	if err != nil {
 		return nil, err
 	}
-
-	tobsCfg := cfg.(*Config)
 
 	exporter, err := exporterhelper.NewMetricsExporter(
 		cfg,
@@ -101,12 +102,8 @@ func createMetricsExporter(
 	if err != nil {
 		return nil, err
 	}
-	ourConfig, ok := cfg.(*Config)
-	if !ok {
-		return nil, fmt.Errorf("invalid config: %#v", cfg)
-	}
 	return resourcetotelemetry.WrapMetricsExporter(
-		ourConfig.Metrics.ResourceAttributes,
+		tobsCfg.Metrics.ResourceAttributes,
 		exporter,
 	), nil
 }