+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - `googlecloudexporter`: fix the `exporter.googlecloud.OTLPDirect` fature-gate, which was not applied when the flag was provided (#9116)
 - `signalfxexporter`: Fix bug to enable timeouts for correlating traces and metrics (#9101)
 - `windowsperfcountersreceiver`: fix exported values being integers instead of doubles (#9138)
+- `tanzuobservabilityexporter`: Make metrics stanza in config be optional (#9098)
 
 ### 🚩 Deprecations 🚩